Cleaning up Dockerfile, adding one for Mongo
I had problems with the provided Dockerfile, so I reworked it, tested it
out, and switched it to using a "builder pattern" instead of trying to
clean up after doing the build.

The "Dockerfile" now builds in the "golang:alpine" Docker image, and
builds an image using the base "alpine:latest" image for a smaller
resulting image.

I also included a "Dockerfile-mongodb" that similarly uses the "builder
pattern" to build based on a MongoDB base Docker image, and then sets
itself up to run MailHog with MongoDB.